Improved Recusrive Doubling Module with DMA
Please refer to the commit at the same time stamp on ICENET to see
detailed changes made to the modules. Commits here are for the configs,
harness and the tester. Tester is derived from the original Recursive
Doubling tester, but modified to have different collective IDs for each
run and only run 24 tests which covers every possible combination of
packets sent for a 8 node system.